backend/requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
python-multipart
boto3
pytesseract
pdf2image
openai>=1.0.0
python-dotenv
aiofiles
chromadb>=0.5.3
numpy<2.0
moviepy
openai-whisper
httpx